requirements.txt
Django==4.2.29
Pillow==9.5.0
python-decouple==3.8
whitenoise==6.7.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-ckeditor==6.7.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
